requirements.txt
altair==5.2.0
asgiref==3.8.0
asttokens==2.4.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crispy-bootstrap4==2024.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
Django==5.0.3
django-crispy-forms==2.2
docopt==0.6.2
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.13.1
Flask==3.0.0
gitdb==4.0.11
GitPython==3.1.40
idna==3.6
importlib-metadata==6.11.0
ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keyboard==0.13.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
numpy==1.23.5
packaging==23.2
pandas==1.5.2
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
Pillow==10.1.0
pipreqs==0.5.0
platformdirs==4.2.0
plotly==5.24.1
prompt_toolkit==3.0.48
protobuf==4.25.1
pure_eval==0.2.3
pyarrow==14.0.1
pydeck==0.8.1b0
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2022.6
pywin32==307
pyzmq==26.2.0
referencing==0.31.1
requests==2.31.0
rich==13.7.0
rpds-py==0.13.2
six==1.16.0
smmap==5.0.1
soupsieve==2.6
sqlparse==0.4.4
stack-data==0.6.3
streamlit==1.29.0
style==1.1.0
tenacity==8.2.3
tinycss2==1.3.0
toml==0.10.2
toolz==0.12.0
tornado==6.4
traitlets==5.14.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
update==0.0.1
urllib3==2.1.0
validators==0.22.0
virtualenv==20.25.1
watchdog==3.0.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.1
yarg==0.1.9
zipp==3.17.0
